Extract RecentFilesService
