+++ packages/host/app/resources/file.ts
@@ -7,6 +7,7 @@ import { logger } from '@cardstack/runtime-common';
 import LoaderService from '../services/loader-service';
 import type MessageService from '../services/message-service';
 import type CardService from '@cardstack/host/services/card-service';
+import type RecentFilesService from '@cardstack/host/services/recent-files-service';
 
 const log = logger('resource:file');
 
@@ -55,6 +56,7 @@ class _FileResource extends Resource<Args> {
   @service declare loaderService: LoaderService;
   @service declare messageService: MessageService;
   @service declare cardService: CardService;
+  @service declare recentFilesService: RecentFilesService;
 
   constructor(owner: unknown) {
     super(owner);
@@ -100,7 +102,7 @@ class _FileResource extends Resource<Args> {
     }
 
     if (newState.state === 'ready') {
-      this.cardService.addRecentFile(newState.url);
+      this.recentFilesService.addRecentFile(newState.url);
     }
   }
